Adjust random values up to the correct max
Previously, the initial maximum was assumed to still be the maximum sent count. This is problematic, because if the last weightedUser is chosen, there's no user left that corresponds to the maximum random choice. This fixes that by adjusting dynamically to the last user's weight, not to the total weight calculated.